Avoid using a static value to seed rng
This probably makes little difference within the context of this
particular demo but this will cause duplicate uuids to be generated
between invocations.

I noticed this while I was working on the Rockset demo, as subsequent
invocations of the generator (restarting `flowctl develop`) would add
events to a collection which had the same uuid key.

This was rather unexpected...